test: tidy up testing constants
